+++ packages/kitchensink-app/storybook/stories/index.ts
@@ -13,5 +13,7 @@ import './Progress/ProgressBar.stories';
 import './RadioButton/RadioButton.stories';
 import './Switch/Switch.stories';
 import './Text/Text.stories';
+import './BottomNavigator/BottomNavigator.stories';
+import './Avatar/Avatar.stories';
 import './Select/Select.stories';
 import './SelectDate/SelectDate.stories';