+++ packages/react-web-kit/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@0.5.3...@tecsinapse/react-web-kit@0.6.0) (2021-07-09)
+
+
+### Bug Fixes
+
+* drawer story ([d92dbe7](https://github.com/tecsinapse/design-system/commit/d92dbe7b4e296ccd2cdb5e99dcd79023cc5c9e63))
+* error build ([161016b](https://github.com/tecsinapse/design-system/commit/161016bb8ac15018e3ca073aa841cb92db594273))
+* minor changes ([851c40a](https://github.com/tecsinapse/design-system/commit/851c40a65fb0b2d1416d61d656881915253f4996))
+* styles in Stories and Drawer. ([18a15ee](https://github.com/tecsinapse/design-system/commit/18a15ee5ff437da760062a58f4ef3314da53ac60))
+* type anchorPosition ([be815cf](https://github.com/tecsinapse/design-system/commit/be815cfb0c73a42628be5ebe468d152ea6e6d9bc))
+
+
+### Features
+
+* add component Drawer ([844f5eb](https://github.com/tecsinapse/design-system/commit/844f5eb193f08a7008782be544f7816254358de4))
+* add styles in MenuBar container items ([827bf23](https://github.com/tecsinapse/design-system/commit/827bf232de984b952015f78a1b1d9e3bccd046cd))
+
+
+
+
+
+## [0.5.3](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@0.5.2...@tecsinapse/react-web-kit@0.5.3) (2021-07-09)
+
+
+### Bug Fixes
+
+* add start command to watch changes using babel (packages not refreshing) ([b038df0](https://github.com/tecsinapse/design-system/commit/b038df008a56f2299e3bbc321f662cad19c1b33d))
+
+
+
+
+
 ## [0.5.2](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-web-kit@0.5.1...@tecsinapse/react-web-kit@0.5.2) (2021-06-18)
 
 **Note:** Version bump only for package @tecsinapse/react-web-kit